2002-05-28  Roland McGrath  <roland@frob.com>
	* pids.c (add_fn_pids): unsigned -> size_t
	* w.c (add_utmp_procs): Likewise.
	* login.c (add_canonical_host): Likewise.
	* ps.c (main): Likewise.
	* login.c (add_entry, main): Likewise.
	* settrans.c (main): int -> size_t
	* showtrans.c (main): Likewise.
	* fsysopts.c (main): Likewise.
	* vmstat.c (main): Use int for FWIDTHS.
	* ping.c (main): size_t -> socklen_t
	* fakeauth.c (S_auth_getids, S_auth_makeauth,
	S_auth_server_authenticate): u_int -> size_t
	* showtrans.c (main): Cast field width/precision args to int.

+++ utils/settrans.c
@@ -90,7 +90,7 @@ main(int argc, char *argv[])
 
   /* The translator's arg vector, in '\0' separated format.  */
   char *argz = 0;
-  int argz_len = 0;
+  size_t argz_len = 0;
 
   /* The control port for any active translator we start up.  */
   fsys_t active_control = MACH_PORT_NULL;
